requirements.txt
pandas==1.2.3
tqdm==4.58.0
wget==3.2
tensorflow==2.4.1
tensorflow-hub==0.11.0
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.1
plotly==4.14.3
psutil==5.8.0